move AWSHeader to impl, restore DefaultS3ClientFactory indentation
